requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
Brotli==1.1.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
curl_cffi==0.13.0
decorator==4.4.2
distro==1.9.0
easyocr==1.7.2
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
frozenlist==1.8.0
fsspec==2025.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
lazy_loader==0.4
llvmlite==0.45.1
MarkupSafe==3.0.3
more-itertools==10.8.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.7.0
mutagen==1.47.0
networkx==3.5
ninja==1.13.0
numba==0.62.1
numpy==2.3.4
openai==2.6.1
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
proglog==0.1.12
propcache==0.4.1
pyclipper==1.3.0.post6
pycparser==2.23
pycryptodomex==3.23.0
pycurl==7.45.7
pydantic==2.12.3
pydantic_core==2.41.4
pytesseract==0.3.13
python-bidi==0.6.6
python-dotenv==1.1.1
pytube==15.0.0
PyYAML==6.0.3
regex==2025.10.22
requests==2.32.5
scikit-image==0.25.2
scipy==1.16.2
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sympy==1.14.0
tifffile==2025.10.16
tiktoken==0.12.0
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.22.0
yt-dlp==2025.10.22
